Fixes for Sync
