+++ msticpy/aiagents/__init__.py
@@ -1,1 +1,6 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
 """Modules related to AI agents used in MSTICpy."""
