+++ msticpy/aiagents/assistant_agent.py
@@ -1,3 +1,8 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
 # pylint: disable=too-few-public-methods
 """Implements the assistant agent for MSTICpy."""
 from autogen.agentchat.contrib.retrieve_assistant_agent import RetrieveAssistantAgent
