+++ msticpy/aiagents/config_utils.py
@@ -1,3 +1,8 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
 """Utility Modules related to AI agents used in MSTICpy."""
 
 import os
